+++ .storybook/decorators/GeneratedStory/README.md
@@ -1,8 +1,10 @@
 # Generated Story Decorator
 
-Chromatic & Storybook provide no way to programmatically define stories for multiple variants. For example, to test the `Button` component with every combination of variant, size, darkMode, and icon position, we would have to explicitly write each of these stories in the `*.stories.tsx file`, resulting in a lot of manual overhead work.
+## Motivation
 
-This custom decorator reads metadata from the storybook default export, and generates a single story that renders all combinations of props defined in the story metadata.
+Storybook & Chromatic provide no way to programmatically define stories for multiple variants. For example, to test the `Button` component with every combination of variant, size, darkMode, icon position etc., we would have to explicitly write each of these stories in the `*.stories.tsx file`, resulting in a lot of manual overhead work. In addition, testing each of these individually would result in hundreds, or even thousands, of stories which would dramatically increase the cost of Chromatic, and generates noise on the sidebar.
+
+As an alternative, this custom decorator reads metadata from the storybook default export, and generates a single story that renders all combinations of props defined in the story metadata.
 
 # Getting started
