+++ .storybook/decorators/GeneratedStory/utils/valStr.ts
@@ -1,0 +1,18 @@
+/**
+ * @returns the provided value as a string
+ */
+export function valStr(val: any): string {
+  const MAX_STR_LEN = 24;
+  if (typeof val === 'object') {
+    if (val.type) {
+      if (typeof val.type === 'string') return `<${val.type} />`;
+      return `<${val.type.displayName} />` ?? 'JSX Element';
+    }
+    if (Array.isArray(val)) return 'Array';
+    else return 'Object';
+  }
+  if (typeof val === 'string' && val.length > MAX_STR_LEN) {
+    return val.slice(0, MAX_STR_LEN) + '…';
+  }
+  return `${val}`;
+}
